Fix CQL binary protocol v3 framing
There's a small but significant difference in the CQL binary protocol
framing in v3 and later -- the stream field is 16 bits, not 8 bits like
in earlier versions.

+++ src/client.rs
@@ -32,7 +32,7 @@ impl Client {
         Client {socket: socket, version: version, prepared: BTreeMap::new()}
     }
 
-    fn build_auth<'a>(&self, creds: &'a Vec<CowStr>, stream: i8) -> CqlRequest<'a> {
+    fn build_auth<'a>(&self, creds: &'a Vec<CowStr>, stream: i16) -> CqlRequest<'a> {
         return CqlRequest {
             version: self.version,
             flags: 0x00,
